requirements.txt
fastapi
uvicorn[standard]
watchfiles
aiofiles
python-decouple
jinja2
aiohttp